refactor: extract TomlSource from staert.go
